requirements.txt
streamlit
tensorflow-cpu==2.12.0
numpy
pillow
protobuf==3.20.3
